feat: add reporting
Add a new window that can be opened from the config that allows
reporting messages and listings that NoSoliciting either didn't filter
but should have or did filter but shouldn't have.

This comes with the added benefit of showing why (both in the
reporting window and the log) a message was filtered for easier
debugging.